Struct re_query_cache::cache::CacheBucket

pub struct CacheBucket {
    pub(crate) data_times: VecDeque<(TimeInt, RowId)>,
    pub(crate) pov_instance_keys: FlatVecDeque<InstanceKey>,
    pub(crate) components: BTreeMap<ComponentName, Box<dyn ErasedFlatVecDeque + Send + Sync>>,
    pub(crate) total_size_bytes: u64,
}

Caches the results of any query for an arbitrary range of time.

This caches all the steps involved in getting data ready for space views:

index search,

instance key joining,

deserialization.

We share the CacheBucket implementation between all types of queries to avoid duplication of logic, especially for things that require metaprogramming, to keep the macro madness to a minimum. In the case of LatestAt queries, a CacheBucket will always contain a single timestamp worth of data.

Fields
data_times: VecDeque<(TimeInt, RowId)>

The data timestamps and RowIds of all cached rows.

This corresponds to the data time and RowId returned by re_query::query_archetype.

This is guaranteed to always be sorted and dense (i.e. there cannot be a hole in the cached data, unless the raw data itself in the store has a hole at that particular point in time).

Reminder: within a single timestamp, rows are sorted according to their RowIds.

Invariant: data_times.len() == pov_instance_keys.num_entries()

pov_instance_keys: FlatVecDeque<InstanceKey>

The InstanceKeys of the point-of-view components.

Invariant: data_times.len() == pov_instance_keys.num_entries()

components: BTreeMap<ComponentName, Box<dyn ErasedFlatVecDeque + Send + Sync>>

The resulting component data, pre-deserialized, pre-joined.

All the contained FlatVecDeques have the same length as data_times.

total_size_bytes: u64

The total size in bytes stored in this bucket.

Only used so we can decrement the global cache size when the last reference to a bucket gets dropped.

Implementations

impl CacheBucket

fn sanity_check(&self)

pub fn time_range(&self) -> Option<TimeRange>

pub fn contains_data_time(&self, data_time: TimeInt) -> bool

pub fn contains_data_row(&self, data_time: TimeInt, row_id: RowId) -> bool

pub fn num_entries(&self) -> usize

How many timestamps’ worth of data is stored in this bucket?

pub fn is_empty(&self) -> bool

pub fn iter_data_times(&self) -> impl Iterator<Item = &(TimeInt, RowId)>

Iterate over the timestamps of the point-of-view components.

pub fn iter_pov_instance_keys(&self) -> impl Iterator<Item = &[InstanceKey]>

Iterate over the InstanceKey batches of the point-of-view components.

pub fn iter_component<C: Component + Send + Sync + 'static>( &self ) -> Option<impl Iterator<Item = &[C]>>

Iterate over the batches of the specified non-optional component.

pub fn iter_component_opt<C: Component + Send + Sync + 'static>( &self ) -> Option<impl Iterator<Item = &[Option<C>]>>

Iterate over the batches of the specified optional component.

pub fn entry_range(&self, time_range: TimeRange) -> Range<usize>

Returns the index range that corresponds to the specified time_range.

Use the returned range with one of the range iteration methods:

Self::range_data_times

Self::range_pov_instance_keys

Self::range_component

Self::range_component_opt

Make sure that the bucket hasn’t been modified in-between!

This is O(2*log(n)), so make sure to clone the returned range rather than calling this multiple times.

pub fn range_data_times( &self, entry_range: Range<usize> ) -> impl Iterator<Item = &(TimeInt, RowId)>

Range over the timestamps of the point-of-view components.

pub fn range_pov_instance_keys( &self, entry_range: Range<usize> ) -> impl Iterator<Item = &[InstanceKey]>

Range over the InstanceKey batches of the point-of-view components.

pub fn component<C: Component + Send + Sync + 'static>( &self ) -> Option<&FlatVecDeque<C>>

Get the raw batches for the specified non-optional component.

pub fn range_component<C: Component + Send + Sync + 'static>( &self, entry_range: Range<usize> ) -> Option<impl Iterator<Item = &[C]>>

Range over the batches of the specified non-optional component.

pub fn component_opt<C: Component + Send + Sync + 'static>( &self ) -> Option<&FlatVecDeque<Option<C>>>

Get the raw batches for the specified optional component.

pub fn range_component_opt<C: Component + Send + Sync + 'static>( &self, entry_range: Range<usize> ) -> Option<impl Iterator<Item = &[Option<C>]>>

Range over the batches of the specified optional component.

pub fn truncate_at_time(&mut self, threshold: TimeInt) -> u64

Removes everything from the bucket that corresponds to a time equal or greater than the specified threshold.

Returns the number of bytes removed.

impl CacheBucket

fn insert_pov1<A, R1>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static,

Alias for Self::insert_pov1_comp0.

pub fn insert_pov1_comp0<A, R1>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 0 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp1<A, R1, C1>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 1 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp2<A, R1, C1, C2>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 2 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp3<A, R1, C1, C2, C3>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 3 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp4<A, R1, C1, C2, C3, C4>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 4 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp5<A, R1, C1, C2, C3, C4, C5>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 5 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp6<A, R1, C1, C2, C3, C4, C5, C6>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 6 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp7<A, R1, C1, C2, C3, C4, C5, C6, C7>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 7 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp8<A, R1, C1, C2, C3, C4, C5, C6, C7, C8>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 8 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

pub fn insert_pov1_comp9<A, R1, C1, C2, C3, C4, C5, C6, C7, C8, C9>( &mut self, query_time: TimeInt, arch_view: &ArchetypeView<A> ) -> Result<u64>where A: Archetype, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, C9: Component + Send + Sync + 'static,

Inserts the contents of the given ArchetypeView, which are made of the specified 1 point-of-view components and 9 optional components, to the cache.

Returns the size in bytes of the data that was cached.

query_time must be the time of query, not of the resulting data.

fn insert_component<A: Archetype, C: Component + Send + Sync + 'static>( &mut self, at: usize, arch_view: &ArchetypeView<A> ) -> Result<u64>

fn insert_component_opt<A: Archetype, C: Component + Send + Sync + 'static>( &mut self, at: usize, arch_view: &ArchetypeView<A> ) -> Result<u64>

This will insert an empty slice for a missing component (instead of N None values).