Struct re_query_cache::cache::Caches

pub struct Caches {
    store_id: StoreId,
    per_cache_key: RwLock<HashMap<CacheKey, Arc<RwLock<CachesPerArchetype>>>>,
}

Maintains the top-level cache mappings.

Fields
store_id: StoreId

The StoreId of the associated [DataStore].

per_cache_key: RwLock<HashMap<CacheKey, Arc<RwLock<CachesPerArchetype>>>>
Implementations

impl Caches

pub fn new(store: &DataStore) -> Self

impl Caches

pub fn clear(&self)

Clears all caches.

pub fn with_latest_at<A, F1, F2, R1, R2>( &self, store: &DataStore, entity_path: EntityPath, query: &LatestAtQuery, upsert: F1, iter: F2 ) -> (Option<R1>, R2)where A: Archetype, F1: FnMut(&mut LatestAtCache) -> R1, F2: FnMut(&LatestAtCache) -> R2,

Gives access to the appropriate LatestAtCache according to the specified query parameters.

upsert is a user-defined callback that will be run first, with full mutable access to the cache. iter is a user-defined callback that will be run last, with shared access.

These callback semantics allow for reentrancy: you can use the same cache from multiple query contexts (i.e. space views), even in a work-stealing environment.

pub fn with_range<A, F1, F2, R1, R2>( &self, store: &DataStore, entity_path: EntityPath, query: &RangeQuery, upsert: F1, iter: F2 ) -> (Option<R1>, R2)where A: Archetype, F1: FnMut(&mut RangeCache) -> R1, F2: FnMut(&RangeCache) -> R2,

Gives access to the appropriate RangeCache according to the specified query parameters.

upsert is a user-defined callback that will be run first, with full mutable access to the cache. iter is a user-defined callback that will be run last, with shared access.

These callback semantics allow for reentrancy: you can use the same cache from multiple query contexts (i.e. space views), even in a work-stealing environment.

impl Caches

pub fn stats(&self, detailed_stats: bool) -> CachesStats

Computes the stats for all primary caches.

per_component toggles per-component stats.

impl Caches

pub fn query_archetype_latest_at_pov1_comp0<'a, A, R1, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1])),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 0 optional components.

pub fn query_archetype_latest_at_pov1_comp1<'a, A, R1, C1, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 1 optional components.

pub fn query_archetype_latest_at_pov1_comp2<'a, A, R1, C1, C2, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 2 optional components.

pub fn query_archetype_latest_at_pov1_comp3<'a, A, R1, C1, C2, C3, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 3 optional components.

pub fn query_archetype_latest_at_pov1_comp4<'a, A, R1, C1, C2, C3, C4, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 4 optional components.

pub fn query_archetype_latest_at_pov1_comp5<'a, A, R1, C1, C2, C3, C4, C5, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 5 optional components.

pub fn query_archetype_latest_at_pov1_comp6<'a, A, R1, C1, C2, C3, C4, C5, C6, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 6 optional components.

pub fn query_archetype_latest_at_pov1_comp7<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 7 optional components.

pub fn query_archetype_latest_at_pov1_comp8<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>, Option<&[Option<C8>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 8 optional components.

pub fn query_archetype_latest_at_pov1_comp9<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, C9, F>( &self, store: &'a DataStore, query: &LatestAtQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, C9: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>, Option<&[Option<C8>]>, Option<&[Option<C9>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 9 optional components.

impl Caches

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view component and no optional components.

Alias for Self::query_archetype_pov1_comp0.

pub fn query_archetype_pov1<'a, A, R1, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1])),

impl Caches

pub fn query_archetype_pov1_comp0<'a, A, R1, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1])),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 0 optional components.

pub fn query_archetype_pov1_comp1<'a, A, R1, C1, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 1 optional components.

pub fn query_archetype_pov1_comp2<'a, A, R1, C1, C2, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 2 optional components.

pub fn query_archetype_pov1_comp3<'a, A, R1, C1, C2, C3, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 3 optional components.

pub fn query_archetype_pov1_comp4<'a, A, R1, C1, C2, C3, C4, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 4 optional components.

pub fn query_archetype_pov1_comp5<'a, A, R1, C1, C2, C3, C4, C5, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 5 optional components.

pub fn query_archetype_pov1_comp6<'a, A, R1, C1, C2, C3, C4, C5, C6, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 6 optional components.

pub fn query_archetype_pov1_comp7<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 7 optional components.

pub fn query_archetype_pov1_comp8<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>, Option<&[Option<C8>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 8 optional components.

pub fn query_archetype_pov1_comp9<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, C9, F>( &self, store: &'a DataStore, query: &AnyQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, C9: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>, Option<&[Option<C8>]>, Option<&[Option<C9>]>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 9 optional components.

impl Caches

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view component and no optional components.

Alias for Self::query_archetype_with_history_pov1_comp0.

pub fn query_archetype_with_history_pov1<'a, A, R1, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1])),

impl Caches

pub fn query_archetype_with_history_pov1_comp0<'a, A, R1, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1])),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 0 optional components.

pub fn query_archetype_with_history_pov1_comp1<'a, A, R1, C1, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 1 optional components.

pub fn query_archetype_with_history_pov1_comp2<'a, A, R1, C1, C2, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 2 optional components.

pub fn query_archetype_with_history_pov1_comp3<'a, A, R1, C1, C2, C3, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 3 optional components.

pub fn query_archetype_with_history_pov1_comp4<'a, A, R1, C1, C2, C3, C4, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 4 optional components.

pub fn query_archetype_with_history_pov1_comp5<'a, A, R1, C1, C2, C3, C4, C5, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 5 optional components.

pub fn query_archetype_with_history_pov1_comp6<'a, A, R1, C1, C2, C3, C4, C5, C6, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 6 optional components.

pub fn query_archetype_with_history_pov1_comp7<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 7 optional components.

pub fn query_archetype_with_history_pov1_comp8<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>, Option<&[Option<C8>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 8 optional components.

pub fn query_archetype_with_history_pov1_comp9<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, C9, F>( &self, store: &'a DataStore, timeline: &'a Timeline, time: &'a TimeInt, history: &ExtraQueryHistory, ent_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, C9: Component + Send + Sync + 'static, F: FnMut(((Option<TimeInt>, RowId), &[InstanceKey], &[R1], Option<&[Option<C1>]>, Option<&[Option<C2>]>, Option<&[Option<C3>]>, Option<&[Option<C4>]>, Option<&[Option<C5>]>, Option<&[Option<C6>]>, Option<&[Option<C7>]>, Option<&[Option<C8>]>, Option<&[Option<C9>]>)),

Cached implementation of re_query::query_archetype_with_history for 1 point-of-view components and 9 optional components.

impl Caches

pub fn query_archetype_range_pov1_comp0<'a, A, R1, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 0 optional components.

pub fn query_archetype_range_pov1_comp1<'a, A, R1, C1, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 1 optional components.

pub fn query_archetype_range_pov1_comp2<'a, A, R1, C1, C2, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 2 optional components.

pub fn query_archetype_range_pov1_comp3<'a, A, R1, C1, C2, C3, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 3 optional components.

pub fn query_archetype_range_pov1_comp4<'a, A, R1, C1, C2, C3, C4, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>, Option<&FlatVecDeque<Option<C4>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 4 optional components.

pub fn query_archetype_range_pov1_comp5<'a, A, R1, C1, C2, C3, C4, C5, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>, Option<&FlatVecDeque<Option<C4>>>, Option<&FlatVecDeque<Option<C5>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 5 optional components.

pub fn query_archetype_range_pov1_comp6<'a, A, R1, C1, C2, C3, C4, C5, C6, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>, Option<&FlatVecDeque<Option<C4>>>, Option<&FlatVecDeque<Option<C5>>>, Option<&FlatVecDeque<Option<C6>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 6 optional components.

pub fn query_archetype_range_pov1_comp7<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>, Option<&FlatVecDeque<Option<C4>>>, Option<&FlatVecDeque<Option<C5>>>, Option<&FlatVecDeque<Option<C6>>>, Option<&FlatVecDeque<Option<C7>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 7 optional components.

pub fn query_archetype_range_pov1_comp8<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>, Option<&FlatVecDeque<Option<C4>>>, Option<&FlatVecDeque<Option<C5>>>, Option<&FlatVecDeque<Option<C6>>>, Option<&FlatVecDeque<Option<C7>>>, Option<&FlatVecDeque<Option<C8>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 8 optional components.

pub fn query_archetype_range_pov1_comp9<'a, A, R1, C1, C2, C3, C4, C5, C6, C7, C8, C9, F>( &self, store: &'a DataStore, query: &RangeQuery, entity_path: &'a EntityPath, f: F ) -> Result<()>where A: Archetype + 'a, R1: Component + Send + Sync + 'static, C1: Component + Send + Sync + 'static, C2: Component + Send + Sync + 'static, C3: Component + Send + Sync + 'static, C4: Component + Send + Sync + 'static, C5: Component + Send + Sync + 'static, C6: Component + Send + Sync + 'static, C7: Component + Send + Sync + 'static, C8: Component + Send + Sync + 'static, C9: Component + Send + Sync + 'static, F: FnMut(bool, Range<usize>, (&VecDeque<(TimeInt, RowId)>, &FlatVecDeque<InstanceKey>, &FlatVecDeque<R1>, Option<&FlatVecDeque<Option<C1>>>, Option<&FlatVecDeque<Option<C2>>>, Option<&FlatVecDeque<Option<C3>>>, Option<&FlatVecDeque<Option<C4>>>, Option<&FlatVecDeque<Option<C5>>>, Option<&FlatVecDeque<Option<C6>>>, Option<&FlatVecDeque<Option<C7>>>, Option<&FlatVecDeque<Option<C8>>>, Option<&FlatVecDeque<Option<C9>>>)),

Cached implementation of re_query::query_archetype and re_query::range_archetype (combined) for 1 point-of-view components and 9 optional components.